*: move hint and block hint processor to util directory (#15897)
